+++ patches/src/main/kotlin/app/revanced/patches/shared/misc/extension/SharedExtensionPatch.kt
@@ -85,8 +85,8 @@ fun sharedExtensionPatch(
 
 class ExtensionHook internal constructor(
     internal val fingerprint: Fingerprint,
-    private val insertIndexResolver: ((Method) -> Int),
-    private val contextRegisterResolver: (Method) -> String,
+    private val insertIndexResolver: BytecodePatchContext.(Method) -> Int,
+    private val contextRegisterResolver: BytecodePatchContext.(Method) -> String,
 ) {
     context(BytecodePatchContext)
     operator fun invoke(extensionClassDescriptor: String) {
@@ -96,20 +96,20 @@ class ExtensionHook internal constructor(
         fingerprint.method.addInstruction(
             insertIndex,
             "invoke-static/range { $contextRegister .. $contextRegister }, " +
-                "$extensionClassDescriptor->setContext(Landroid/content/Context;)V",
+                    "$extensionClassDescriptor->setContext(Landroid/content/Context;)V",
         )
     }
 }
 
 fun extensionHook(
-    insertIndexResolver: ((Method) -> Int) = { 0 },
-    contextRegisterResolver: (Method) -> String = { "p0" },
+    insertIndexResolver: BytecodePatchContext.(Method) -> Int = { 0 },
+    contextRegisterResolver: BytecodePatchContext.(Method) -> String = { "p0" },
     fingerprint: Fingerprint,
 ) = ExtensionHook(fingerprint, insertIndexResolver, contextRegisterResolver)
 
 fun extensionHook(
-    insertIndexResolver: ((Method) -> Int) = { 0 },
-    contextRegisterResolver: (Method) -> String = { "p0" },
+    insertIndexResolver: BytecodePatchContext.(Method) -> Int = { 0 },
+    contextRegisterResolver: BytecodePatchContext.(Method) -> String = { "p0" },
     fingerprintBuilderBlock: FingerprintBuilder.() -> Unit,
 ) = {
     val fingerprint by FingerprintDelegate(block = fingerprintBuilderBlock)