+++ patches/src/main/kotlin/app/revanced/patches/spotify/misc/Fingerprints.kt
@@ -42,13 +42,37 @@ internal val contextMenuViewModelClassFingerprint by fingerprint {
     strings("ContextMenuViewModel(header=")
 }
 
-internal val contextMenuViewModelAddItemFingerprint by fingerprint {
+/**
+ * Used in versions older than "9.0.60.128".
+ */
+internal val oldContextMenuViewModelAddItemFingerprint by fingerprint {
     parameters("L")
     returns("V")
     custom { method, _ ->
         method.indexOfFirstInstruction {
             getReference<MethodReference>()?.name == "add"
         } >= 0
+    }
+}
+
+internal val contextMenuViewModelConstructorFingerprint = fingerprint {
+    accessFlags(AccessFlags.PUBLIC, AccessFlags.CONSTRUCTOR)
+}
+
+/**
+ * Used to find the interface name of a context menu item.
+ */
+internal val browsePodcastsContextMenuItemClassFingerprint = fingerprint {
+    strings("browse_podcast_item", "ui_navigate")
+}
+
+internal const val CONTEXT_MENU_ITEM_CLASS_DESCRIPTOR_PLACEHOLDER = "Lapp/revanced/ContextMenuItemPlaceholder;"
+internal val extensionFilterContextMenuItemsFingerprint = fingerprint {
+    accessFlags(AccessFlags.PUBLIC, AccessFlags.STATIC)
+    returns("Ljava/util/List;")
+    parameters("Ljava/util/List;")
+    custom { method, classDef ->
+        method.name == "filterContextMenuItems" && classDef.type == EXTENSION_CLASS_DESCRIPTOR
     }
 }
 
@@ -93,7 +117,7 @@ internal val abstractProtobufListEnsureIsMutableFingerprint by fingerprint {
     }
 }
 
-private fun structureGetSectionsFingerprint(className: String) = fingerprint {
+internal fun structureGetSectionsFingerprint(className: String) = fingerprint {
     custom { method, classDef ->
         classDef.endsWith(className) && method.indexOfFirstInstruction {
             opcode == Opcode.IGET_OBJECT && getReference<FieldReference>()?.name == "sections_"
@@ -109,7 +133,7 @@ internal val homeStructureGetSectionsFingerprint by
     structureGetSectionsFingerprint("homeapi/proto/HomeStructure;")
 
 internal val browseSectionFingerprint by fingerprint {
-    custom { _, classDef -> classDef.endsWith("browsita/v1/resolved/Section;") }
+    custom { _, classDef-> classDef.endsWith("browsita/v1/resolved/Section;") }
 }
 
 internal val browseStructureGetSectionsFingerprint by