+++ patches/src/main/kotlin/app/revanced/patches/spotify/misc/fix/Fingerprints.kt
@@ -2,8 +2,10 @@ package app.revanced.patches.spotify.misc.fix
 
 import app.revanced.patcher.fingerprint
 import app.revanced.patcher.string
+import app.revanced.util.getReference
+import app.revanced.util.indexOfFirstInstruction
 import com.android.tools.smali.dexlib2.AccessFlags
-import org.stringtemplate.v4.compiler.Bytecode.instructions
+import com.android.tools.smali.dexlib2.iface.reference.MethodReference
 
 internal val getPackageInfoFingerprint by fingerprint {
     instructions(
@@ -11,7 +13,7 @@ internal val getPackageInfoFingerprint by fingerprint {
     )
 }
 
-internal val startLiborbitFingerprint by fingerprint {
+internal val loadOrbitLibraryFingerprint by fingerprint {
     strings("/liborbit-jni-spotify.so")
 }
 
@@ -22,10 +24,38 @@ internal val startupPageLayoutInflateFingerprint by fingerprint {
     strings("blueprintContainer", "gradient", "valuePropositionTextView")
 }
 
-internal val standardIntegrityTokenProviderBuilderFingerprint by fingerprint {
-    strings(
-        "standard_pi_init",
-        "outcome",
-        "success"
+internal val renderStartLoginScreenFingerprint by fingerprint {
+    strings("authenticationButtonFactory", "MORE_OPTIONS")
+}
+
+internal val renderSecondLoginScreenFingerprint by fingerprint {
+    strings("authenticationButtonFactory", "intent_login")
+}
+
+internal val renderThirdLoginScreenFingerprint = fingerprint {
+    strings("EMAIL_OR_USERNAME", "listener")
+}
+
+internal val thirdLoginScreenLoginOnClickFingerprint = fingerprint {
+    strings("login", "listener", "none")
+}
+
+internal val runIntegrityVerificationFingerprint = fingerprint {
+    accessFlags(AccessFlags.PUBLIC, AccessFlags.FINAL)
+    returns("V")
+    opcodes(
+        Opcode.CHECK_CAST,
+        Opcode.INVOKE_VIRTUAL,
+        Opcode.INVOKE_STATIC, // Calendar.getInstance()
+        Opcode.MOVE_RESULT_OBJECT,
+        Opcode.INVOKE_VIRTUAL, // instance.get(6)
+        Opcode.MOVE_RESULT,
+        Opcode.IF_EQ, // if (x == instance.get(6)) return
     )
+    custom { method, _ ->
+        method.indexOfFirstInstruction {
+            val reference = getReference<MethodReference>()
+            reference?.definingClass == "Ljava/util/Calendar;" && reference.name == "get"
+        } >= 0
+    }
 }