+++ patches/src/main/kotlin/app/revanced/patches/spotify/misc/fix/SpoofClientPatch.kt
@@ -9,11 +9,9 @@ import app.revanced.patcher.patch.intOption
 import app.revanced.patches.shared.misc.hex.HexPatchBuilder
 import app.revanced.patches.shared.misc.hex.hexPatch
 import app.revanced.patches.spotify.misc.extension.sharedExtensionPatch
-import app.revanced.util.findInstructionIndicesReversedOrThrow
-import app.revanced.util.getReference
-import app.revanced.util.indexOfFirstInstructionReversedOrThrow
-import app.revanced.util.returnEarly
+import app.revanced.util.*
 import com.android.tools.smali.dexlib2.Opcode
+import com.android.tools.smali.dexlib2.iface.instruction.FiveRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.reference.MethodReference
 
@@ -24,8 +22,8 @@ val spoofClientPatch = bytecodePatch(
     name = "Spoof client",
     description = "Spoofs the client to fix various functions of the app.",
 ) {
-    val port by intOption(
-        key = "port",
+    val requestListenerPort by intOption(
+        key = "requestListenerPort",
         default = 4345,
         title = " Login request listener port",
         description = "The port to use for the listener that intercepts and handles login requests. " +
@@ -46,10 +44,10 @@ val spoofClientPatch = bytecodePatch(
                 "x86",
                 "x86_64"
             ).forEach { architecture ->
-                "https://login5.spotify.com/v3/login" to "http://127.0.0.1:$port/v3/login" inFile
+                "https://login5.spotify.com/v3/login" to "http://127.0.0.1:$requestListenerPort/v3/login" inFile
                         "lib/$architecture/liborbit-jni-spotify.so"
 
-                "https://login5.spotify.com/v4/login" to "http://127.0.0.1:$port/v4/login" inFile
+                "https://login5.spotify.com/v4/login" to "http://127.0.0.1:$requestListenerPort/v4/login" inFile
                         "lib/$architecture/liborbit-jni-spotify.so"
             }
         })
@@ -58,6 +56,8 @@ val spoofClientPatch = bytecodePatch(
     compatibleWith("com.spotify.music")
 
     execute {
+        // region Spoof package info.
+
         getPackageInfoFingerprint.method.apply {
             // region Spoof signature.
 
@@ -99,28 +99,93 @@ val spoofClientPatch = bytecodePatch(
             // endregion
         }
 
-        startLiborbitFingerprint.method.addInstructions(
+        // endregion
+
+        // region Spoof client.
+
+        loadOrbitLibraryFingerprint.method.addInstructions(
             0,
             """
-                const/16 v0, $port
-                invoke-static { v0 }, $EXTENSION_CLASS_DESCRIPTOR->listen(I)V
+                const/16 v0, $requestListenerPort
+                invoke-static { v0 }, $EXTENSION_CLASS_DESCRIPTOR->launchListener(I)V
             """
         )
 
         startupPageLayoutInflateFingerprint.method.apply {
             val openLoginWebViewDescriptor =
-                "$EXTENSION_CLASS_DESCRIPTOR->login(Landroid/view/LayoutInflater;)V"
+                "$EXTENSION_CLASS_DESCRIPTOR->launchLogin(Landroid/view/LayoutInflater;)V"
 
             addInstructions(
                 0,
+                "invoke-static/range { p1 .. p1 }, $openLoginWebViewDescriptor"
+            )
+        }
+
+        renderStartLoginScreenFingerprint.method.apply {
+            val onEventIndex = indexOfFirstInstructionOrThrow {
+                opcode == Opcode.INVOKE_INTERFACE && getReference<MethodReference>()?.name == "getView"
+            }
+
+            val buttonRegister = getInstruction<OneRegisterInstruction>(onEventIndex + 1).registerA
+
+            addInstruction(
+                onEventIndex + 2,
+                "invoke-static { v$buttonRegister }, $EXTENSION_CLASS_DESCRIPTOR->setNativeLoginHandler(Landroid/view/View;)V"
+            )
+        }
+
+        renderSecondLoginScreenFingerprint.method.apply {
+            val getViewIndex = indexOfFirstInstructionOrThrow {
+                opcode == Opcode.INVOKE_INTERFACE && getReference<MethodReference>()?.name == "getView"
+            }
+
+            val buttonRegister = getInstruction<OneRegisterInstruction>(getViewIndex + 1).registerA
+
+            // Early return the render for loop since the first item of the loop is the login button.
+            addInstructions(
+                getViewIndex + 2,
                 """
-                    move-object/from16 v3, p1
-                    invoke-static { v3 }, $openLoginWebViewDescriptor
+                    invoke-virtual { v$buttonRegister }, Landroid/view/View;->performClick()Z
+                    return-void
                 """
             )
         }
 
+        renderThirdLoginScreenFingerprint.method.apply {
+            val invokeSetListenerIndex = indexOfFirstInstructionOrThrow {
+                val reference = getReference<MethodReference>()
+                reference?.definingClass == "Landroid/view/View;" && reference.name == "setOnClickListener"
+            }
+
+            val buttonRegister = getInstruction<FiveRegisterInstruction>(invokeSetListenerIndex).registerC
+
+            addInstruction(
+                invokeSetListenerIndex + 1,
+                "invoke-virtual { v$buttonRegister }, Landroid/view/View;->performClick()Z"
+            )
+        }
+
+        thirdLoginScreenLoginOnClickFingerprint.method.apply {
+            // Use placeholder credentials to pass the login screen.
+            val loginActionIndex = indexOfFirstInstructionOrThrow(Opcode.RETURN_VOID) - 1
+            val loginActionInstruction = getInstruction<FiveRegisterInstruction>(loginActionIndex)
+
+            addInstructions(
+                loginActionIndex,
+                """
+                    const-string v${loginActionInstruction.registerD}, "placeholder"
+                    const-string v${loginActionInstruction.registerE}, "placeholder"
+                """
+            )
+        }
+
+        // endregion
+
+        // region Disable verdicts.
+
         // Early return to block sending bad verdicts to the API.
-        standardIntegrityTokenProviderBuilderFingerprint.method.returnEarly()
+        runIntegrityVerificationFingerprint.method.returnEarly()
+
+        // endregion
     }
 }